Add `version` command to new subctl lib
