+++ pkg/subctl/cmd/utils/utils.go
@@ -20,9 +20,8 @@ package utils
 
 import (
 	"fmt"
+	"github.com/submariner-io/submariner-operator/pkg/version"
 	"os"
-
-	"github.com/submariner-io/submariner-operator/pkg/subctl/cmd/version"
 )
 
 // PanicOnError will print the subctl version and then panic in case of an actual error
